Add files via upload
Made some slight changes to the assumptions since I did not include the fact that the material is homogeneous and the strain measurement direction is in the same direction as the magnetic field. I also made a slight math error by not including the phase in the equation for theta.